requirements.txt
asgiref==3.7.2
boto3==1.28.79
botocore==1.31.79
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.7
djangorestframework==3.14.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
